Refactor AspNetCore project
