+++ RiotSharp.AspNetCore/Caching/DistributedCache.cs
@@ -1,25 +1,25 @@
-﻿using Microsoft.Extensions.Caching.Memory;
+﻿using Microsoft.Extensions.Caching.Distributed;
 using System;
 using System.Collections.Generic;
 using RiotSharp.Caching;
 
-namespace RiotSharp.AspNetCore
+namespace RiotSharp.AspNetCore.Caching
 {
     /// <summary>
-    /// Implementation of <see cref="ICache"/> with AspNetCore's local in-memory cache
+    /// Implementation of ICache with AspNetCore's distributed cache
     /// </summary>
-    public class MemoryCache : ICache
+    public class DistributedCache : ICache
     {
-        private readonly IMemoryCache _memoryCache;
+        private readonly IDistributedCache _distributed;
         private readonly List<object> _usedKeys;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="MemoryCache"/> class.
+        /// Initializes a new instance of the <see cref="DistributedCache"/> class.
         /// </summary>
         /// <param name="memoryCache">The memory cache.</param>
-        public MemoryCache(IMemoryCache memoryCache)
+        public DistributedCache(IDistributedCache memoryCache)
         {
-            _memoryCache = memoryCache;
+            _distributed = memoryCache;
             _usedKeys = new List<object>();
         }
 
@@ -27,22 +27,22 @@ namespace RiotSharp.AspNetCore
         public void Add<TK, TV>(TK key, TV value, TimeSpan slidingExpiry) where TV : class
         {
             _usedKeys.Add(key);
-            _memoryCache.Set(key, value, slidingExpiry);
+            _distributed.SetJson(key.ToString(), value, slidingExpiry);
         }
 
         /// <inheritdoc />
         public void Add<TK, TV>(TK key, TV value, DateTime absoluteExpiry) where TV : class
         {
             _usedKeys.Add(key);
-            _memoryCache.Set(key, value, absoluteExpiry);
+            _distributed.SetJson(key.ToString(), value, absoluteExpiry);
         }
 
         /// <inheritdoc />
         public void Clear()
         {
             foreach (var usedKey in _usedKeys)
-            {    
-                _memoryCache.Remove(usedKey);
+            {
+                _distributed.Remove(usedKey.ToString());
                 _usedKeys.Remove(usedKey);
             }
         }
@@ -50,13 +50,13 @@ namespace RiotSharp.AspNetCore
         /// <inheritdoc />
         public TV Get<TK, TV>(TK key) where TV : class
         {
-            _memoryCache.TryGetValue(key, out TV output);
-            return output;
+            return _distributed.GetJson<TV>(key.ToString());
         }
 
+        /// <inheritdoc />
         public void Remove<TK>(TK key)
         {
-            _memoryCache.Remove(key);
+            _distributed.Remove(key.ToString());
         }
     }
 }
